test (LottoTest): LottoTest 작성
